+++ packages/freezed/lib/src/templates/concrete_template.dart
@@ -1,6 +1,7 @@
 import 'package:analyzer/dart/element/element.dart';
 import 'package:freezed/src/models.dart';
 import 'package:freezed/src/templates/properties.dart';
+import 'package:freezed/src/tools/type.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:source_gen/source_gen.dart';
 
@@ -367,11 +368,16 @@ String? parseTypeSource(VariableElement element) {
         element.type.isDynamic &&
         element.type.element!.isSynthetic) {
       final match = RegExp(r'(\w+\??)\s+$').firstMatch(source);
-      type = match?.group(1);
+      return match?.group(1);
     } else if (element.type.element != null) {
       final match = RegExp(r'(\w+<.+?>\??)\s+$').firstMatch(source);
-      type = match?.group(1) ?? type;
+      return match?.group(1) ?? type;
     }
   }
-  return type;
+
+  return resolveFullTypeStringFrom(
+    element.library!,
+    element.type,
+    withNullability: true,
+  );
 }