+++ packages/freezed/CHANGELOG.md
@@ -1,5 +1,6 @@
 [Unreleased]
 
+- Now supports types coming from `import '...' as alias;`
 - generated .freezed.dart are now excluded from code coverage reports.
 - if `toString` is overriden, Freezed no-longer applies `Diagnosticable` which would break the `toString` (fixes #221)